+++ src/util/hamming_dist.c
@@ -9,7 +9,7 @@ int hamming_dist(const void *in_a, const void *in_b, int len)
     const unsigned char *b = in_b;	
 	int d=0;
 
-	for(i=0; i<len; i++)
+	for(i=0; i<(size_t)len; i++)
 	{		
 		x = a[i]^b[i];
 		for (j=0; j<8; j++)
